moved type and point-based heap analyses under impl package
