+++ lisa/src/test/java/it/unive/lisa/cron/heap/TypeBasedHeapTest.java
@@ -6,7 +6,7 @@ import it.unive.lisa.AnalysisSetupException;
 import it.unive.lisa.AnalysisTestExecutor;
 import it.unive.lisa.LiSAConfiguration;
 import it.unive.lisa.analysis.AbstractState;
-import it.unive.lisa.analysis.heap.TypeBasedHeap;
+import it.unive.lisa.analysis.impl.heap.TypeBasedHeap;
 import it.unive.lisa.analysis.impl.numeric.Interval;
 import org.junit.Test;
